requirements.txt
imageio==2.31.6
matplotlib==3.4.3
numpy==1.26.0
Pillow==8.3.2
scipy==1.11.3